feat: move coach UI to shared package
