+++ apps/pronunco/package.json
@@ -12,7 +12,8 @@
     "dexie-react-hooks": "^1.1.7",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-router-dom": "^6.30.1"
+    "react-router-dom": "^6.30.1",
+    "coach-ui": "workspace:*"
   },
   "devDependencies": {
     "@testing-library/react": "^16.3.0",
