+++ apps/pronunco/vitest.config.ts
@@ -11,5 +11,6 @@ export default defineConfig({
     fileParallelism: false,    // serialise files; speed hit is tiny (<200 ms)
     hookTimeout: 10_000,
     setupFiles: ['./tests/setup-vitest.ts'],
+    deps: { inline: ['coach-ui'] },
   },
 });
