+++ apps/sober-body/package.json
@@ -18,7 +18,8 @@
     "react-dom": "^19.1.0",
     "react-router-dom": "^6.30.1",
     "react-select": "^5.10.1",
-    "jszip": "^3.10.1"
+    "jszip": "^3.10.1",
+    "coach-ui": "workspace:*"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
