+++ packages/coach-ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PronunciationCoachUI } from './src/PronunciationCoachUI';
