+++ packages/coach-ui/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "coach-ui",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "dependencies": {
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^6.30.1"
+  }
+}
