+++ packages/coach-ui/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../apps/sober-body/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src"]
+}
